Refactor H5Wasm providers to use same worker-based API
