+++ packages/h5wasm/src/index.ts
@@ -1,3 +1,3 @@
-export { default as H5WasmProvider } from './H5WasmProvider';
-export { default as H5WasmLocalFileProvider } from './local/H5WasmLocalFileProvider';
+export { default as H5WasmBufferProvider } from './H5WasmBufferProvider';
+export { default as H5WasmLocalFileProvider } from './H5WasmLocalFileProvider';
 export { Plugin } from './models';
