+++ packages/h5wasm/src/models.ts
@@ -1,9 +1,3 @@
-import type { Group as H5WasmGroup } from 'h5wasm';
-
-export type H5WasmEntity = ReturnType<H5WasmGroup['get']>;
-
-export type H5WasmAttributes = H5WasmGroup['attrs'];
-
 export interface HDF5Diag {
   major: string;
   minor: string;
