+++ packages/h5wasm/src/utils.ts
@@ -1,43 +1,7 @@
-import {
-  assertDefined,
-  assertNonNull,
-  isNumericType,
-} from '@h5web/shared/guards';
-import type { H5T_STR } from '@h5web/shared/h5t';
-import { H5T_CLASS, H5T_ORDER, H5T_SIGN } from '@h5web/shared/h5t';
-import type {
-  Attribute,
-  ChildEntity,
-  DType,
-  Group,
-  ProvidedEntity,
-  Shape,
-} from '@h5web/shared/hdf5-models';
-import { DTypeClass, EntityKind } from '@h5web/shared/hdf5-models';
-import {
-  arrayType,
-  bitfieldType,
-  buildEntityPath,
-  compoundOrCplxType,
-  enumOrBoolType,
-  floatType,
-  intOrUintType,
-  opaqueType,
-  referenceType,
-  strType,
-  timeType,
-  unknownType,
-} from '@h5web/shared/hdf5-utils';
-import type { Metadata } from 'h5wasm';
-import {
-  BrokenSoftLink as H5WasmSoftLink,
-  Dataset as H5WasmDataset,
-  Datatype as H5WasmDatatype,
-  ExternalLink as H5WasmExternalLink,
-  Group as H5WasmGroup,
-} from 'h5wasm';
+import type { DType } from '@h5web/shared/hdf5-models';
+import { DTypeClass } from '@h5web/shared/hdf5-models';
 
-import type { H5WasmAttributes, H5WasmEntity, HDF5Diag } from './models';
+import type { HDF5Diag } from './models';
 import { Plugin } from './models';
 
 // https://support.hdfgroup.org/services/contributions.html
@@ -52,217 +16,6 @@ export const PLUGINS_BY_FILTER_ID: Record<number, Plugin> = {
   32_017: Plugin.SZ,
   32_026: Plugin.Blosc2,
 };
-
-export function parseEntity(
-  name: string,
-  path: string,
-  h5wEntity: H5WasmEntity,
-  isChild: true,
-): ChildEntity;
-
-export function parseEntity(
-  name: string,
-  path: string,
-  h5wEntity: H5WasmEntity,
-  isChild?: false,
-): ProvidedEntity;
-
-export function parseEntity(
-  name: string,
-  path: string,
-  h5wEntity: H5WasmEntity,
-  isChild = false,
-): ProvidedEntity | ChildEntity {
-  const baseEntity = { name, path };
-
-  if (h5wEntity instanceof H5WasmGroup) {
-    const baseGroup: Group = {
-      ...baseEntity,
-      kind: EntityKind.Group,
-      attributes: parseAttributes(h5wEntity.attrs),
-    };
-
-    if (isChild) {
-      return baseGroup;
-    }
-
-    return {
-      ...baseGroup,
-      children: h5wEntity.keys().map((childName) => {
-        const h5wChild = h5wEntity.get(childName);
-
-        const childPath = buildEntityPath(path, childName);
-        return parseEntity(childName, childPath, h5wChild, true);
-      }),
-    };
-  }
-
-  if (h5wEntity instanceof H5WasmDataset) {
-    const { metadata, filters } = h5wEntity;
-    const { chunks, maxshape, shape, ...rawType } = metadata; // keep `rawType` concise
-
-    return {
-      ...baseEntity,
-      kind: EntityKind.Dataset,
-      shape: metadata.shape,
-      type: parseDType(metadata),
-      chunks: metadata.chunks ?? undefined,
-      filters,
-      attributes: parseAttributes(h5wEntity.attrs),
-      rawType,
-    };
-  }
-
-  if (h5wEntity instanceof H5WasmSoftLink) {
-    return {
-      ...baseEntity,
-      attributes: [],
-      kind: EntityKind.Unresolved,
-      link: {
-        class: 'Soft',
-        path: h5wEntity.target,
-      },
-    };
-  }
-
-  if (h5wEntity instanceof H5WasmExternalLink) {
-    return {
-      ...baseEntity,
-      attributes: [],
-      kind: EntityKind.Unresolved,
-      link: {
-        class: 'External',
-        path: h5wEntity.obj_path,
-        file: h5wEntity.filename,
-      },
-    };
-  }
-
-  return {
-    ...baseEntity,
-    attributes: [],
-    kind:
-      h5wEntity instanceof H5WasmDatatype
-        ? EntityKind.Datatype
-        : EntityKind.Unresolved,
-  };
-}
-
-function parseAttributes(h5wAttrs: H5WasmAttributes): Attribute[] {
-  return Object.entries(h5wAttrs).map(([name, attr]) => {
-    const { shape, metadata } = attr;
-    return {
-      name,
-      shape: shape as Shape,
-      type: parseDType(metadata),
-    };
-  });
-}
-
-// eslint-disable-next-line sonarjs/cognitive-complexity
-export function parseDType(metadata: Metadata): DType {
-  const { type: h5tClass, size } = metadata;
-
-  if (h5tClass === H5T_CLASS.INTEGER) {
-    const { signed, littleEndian } = metadata;
-    return intOrUintType(
-      signed ? H5T_SIGN.SIGN_2 : H5T_SIGN.NONE,
-      size * 8,
-      littleEndian ? H5T_ORDER.LE : H5T_ORDER.BE,
-    );
-  }
-  if (h5tClass === H5T_CLASS.FLOAT) {
-    const { littleEndian } = metadata;
-    return floatType(size * 8, littleEndian ? H5T_ORDER.LE : H5T_ORDER.BE);
-  }
-
-  if (h5tClass === H5T_CLASS.TIME) {
-    return timeType();
-  }
-
-  if (h5tClass === H5T_CLASS.STRING) {
-    const { cset, strpad, vlen } = metadata;
-    return strType(cset, strpad as H5T_STR, vlen ? undefined : size);
-  }
-
-  if (h5tClass === H5T_CLASS.BITFIELD) {
-    return bitfieldType();
-  }
-
-  if (h5tClass === H5T_CLASS.OPAQUE) {
-    return opaqueType();
-  }
-
-  if (h5tClass === H5T_CLASS.COMPOUND) {
-    const { compound_type } = metadata;
-    assertDefined(compound_type);
-
-    return compoundOrCplxType(
-      Object.fromEntries(
-        compound_type.members.map((member) => [
-          member.name,
-          parseDType(member),
-        ]),
-      ),
-    );
-  }
-
-  if (h5tClass === H5T_CLASS.REFERENCE) {
-    return referenceType();
-  }
-
-  if (h5tClass === H5T_CLASS.ENUM) {
-    const { enum_type } = metadata;
-    assertDefined(enum_type);
-    const { members, type } = enum_type;
-
-    const baseType = parseDType({ ...metadata, type });
-    if (!isNumericType(baseType)) {
-      throw new TypeError('Expected enum type to have numeric base type');
-    }
-
-    return enumOrBoolType(baseType, members);
-  }
-
-  if (h5tClass === H5T_CLASS.VLEN) {
-    // Not currently provided, so unable to know base type
-    // const { array_type } = metadata;
-    // assertDefined(array_type);
-    return arrayType(unknownType());
-  }
-
-  if (h5tClass === H5T_CLASS.ARRAY) {
-    const { array_type } = metadata;
-    assertDefined(array_type);
-    assertNonNull(array_type.shape);
-    return arrayType(parseDType(array_type), array_type.shape);
-  }
-
-  return unknownType();
-}
-
-export function readSelectedValue(
-  h5wDataset: H5WasmDataset,
-  selection: string | undefined,
-) {
-  if (!selection) {
-    return h5wDataset.value;
-  }
-
-  const { shape } = h5wDataset;
-  assertNonNull(shape);
-
-  const selectionMembers = selection.split(',');
-  const ranges = selectionMembers.map<[number, number]>((member, i) => {
-    if (member === ':') {
-      return [0, shape[i]];
-    }
-
-    return [Number(member), Number(member) + 1];
-  });
-
-  return h5wDataset.slice(ranges);
-}
 
 export function hasBigInts(type: DType): boolean {
   if (type.class === DTypeClass.Array || type.class === DTypeClass.Enum) {
